requirements.txt
pymupdf
camelot-py[pdf]
python-docx
pandas
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
langchain-core
sentence-transformers
faiss-cpu